refactor(server): move llm_provider_* into providers/ package
