+++ server/src/providers/chatgpt.py
@@ -4,7 +4,7 @@ ChatGPT/OpenAI Provider for metadata generation using OpenAI API
 
 import json
 from typing import Any, override
-from llm_provider_base import (
+from .base import (
     LLMProviderBase,
     EditGenerationRequest,
     EditGenerationResponse,
