+++ server/test/test_normalize_keywords_structure.py
@@ -1,6 +1,6 @@
 import unittest
 
-from llm_provider_base import LLMProviderBase
+from providers.base import LLMProviderBase
 
 
 class _StubProvider(LLMProviderBase):
